Add functions create_array, _strdup, str_concat, and alloc_grid
This commit adds four new functions to the malloc_free directory:
- create_array: creates an array of chars and initializes it with a specific char
- _strdup: returns a pointer to a newly allocated space in memory, which contains a copy of the string given as a parameter
- str_concat: concatenates two strings
- alloc_grid: returns a pointer to a 2-dimensional array of integers

These functions are implemented in separate C files and are accompanied by their respective header file declarations in main.h. The commit also includes a README.md file with a description of the project and its learning objectives.

The commit also includes changes to the 5-strstr.c file in the pointers_arrays_strings directory. The _strstr function has been modified to fix a bug where it would return the haystack string when the needle string was empty.

Lastly, the commit includes a new main.c file in the malloc_free directory, which contains a main function to test the alloc_grid function.

These changes aim to enhance the functionality of the malloc_free project and improve the understanding of dynamic memory allocation in C.